pyproject.toml
[project]
name = "obs-common"
description = "code shared between obs-team repositories"
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
  "click",
  "google-cloud-pubsub",
  "google-cloud-storage",
  "requests",
  "sentry-sdk",
]
dynamic = ["version"]

[project.scripts]
license-check = "obs_common.license_check:main"
service-status = "obs_common.service_status:main"
update-obs-common = "obs_common.update_obs_common:main"
gcs-cli = "obs_common.gcs_cli:main"
pubsub-cli = "obs_common.pubsub_cli:main"
sentry-wrap = "obs_common.sentry_wrap:cli_main"

[build-system]
requires = ["setuptools", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.distutils.bdist_wheel]
universal = true

[tool.setuptools]
packages = ["obs_common"]

[tool.setuptools_scm]
# Empty is fine


[tool.ruff]
line-length = 88
src = ["src", "tests"]
target-version = "py38"

[tool.ruff.lint]
# Enable pycodestyle (E), pyflakes (F), bugbear (B), and bandit (S) rules
select = ["E", "F", "B", "S"]

ignore = ["E501"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"obs_common/license_check.py" = ["S603", "S607"]
"obs_common/release.py" = ["S603", "S607"]
"obs_common/sentry_wrap.py" = ["S603"]
"obs_common/waitfor.py" = ["S310"]
"tests/**/*.py" = ["S101", "S603"]


[tool.pytest.ini_options]
filterwarnings = [
    "error",
]
